新增移动端下载

+++ eam/service/service-eam/src/main/java/com/dt/platform/eam/controller/AssetEmployeeApplyController.java
@@ -2,6 +2,7 @@ package com.dt.platform.eam.controller;
 
 
 import java.util.List;
+import java.util.ArrayList;
 
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -51,7 +52,7 @@ import com.github.foxnic.api.validate.annotations.NotNull;
  * 领用申请 接口控制器
  * </p>
  * @author 金杰 , maillank@qq.com
- * @since 2022-06-23 15:52:11
+ * @since 2022-06-29 19:15:25
 */
 
 @Api(tags = "领用申请")
@@ -68,13 +69,13 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "添加领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.BUSINESS_CODE , value = "业务编号" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.STATUS , value = "办理状态" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class , example = "12"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class , example = "2"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class , example = "1212"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class , example = "0"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.RECORD_TIME , value = "记录时间" , required = false , dataTypeClass=Date.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORIGINATOR_ID , value = "制单人" , required = false , dataTypeClass=String.class),
@@ -95,13 +96,24 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "删除领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class)
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696")
 	})
 	@ApiOperationSupport(order=2)
 	@NotNull(name = AssetEmployeeApplyVOMeta.ID)
 	@SentinelResource(value = AssetEmployeeApplyServiceProxy.DELETE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(AssetEmployeeApplyServiceProxy.DELETE)
 	public Result deleteById(String id) {
+		this.validator().asserts(id).require("缺少id值");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+		// 引用校验
+		Boolean hasRefer = assetEmployeeApplyService.hasRefers(id);
+		// 判断是否可以删除
+		this.validator().asserts(hasRefer).mustInList("不允许删除当前记录",false);
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
 		Result result=assetEmployeeApplyService.deleteByIdLogical(id);
 		return result;
 	}
@@ -120,8 +132,43 @@ public class AssetEmployeeApplyController extends SuperController {
 	@SentinelResource(value = AssetEmployeeApplyServiceProxy.DELETE_BY_IDS , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(AssetEmployeeApplyServiceProxy.DELETE_BY_IDS)
 	public Result deleteByIds(List<String> ids) {
-		Result result=assetEmployeeApplyService.deleteByIdsLogical(ids);
-		return result;
+
+		// 参数校验
+		this.validator().asserts(ids).require("缺少ids参数");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+
+		// 查询引用
+		Map<String, Boolean> hasRefersMap = assetEmployeeApplyService.hasRefers(ids);
+		// 收集可以删除的ID值
+		List<String> canDeleteIds = new ArrayList<>();
+		for (Map.Entry<String, Boolean> e : hasRefersMap.entrySet()) {
+			if (!e.getValue()) {
+				canDeleteIds.add(e.getKey());
+			}
+		}
+
+		// 执行删除
+		if (canDeleteIds.isEmpty()) {
+			// 如果没有一行可以被删除
+			return ErrorDesc.failure().message("无法删除您选中的数据行");
+		} else if (canDeleteIds.size() == ids.size()) {
+			// 如果全部可以删除
+			Result result=assetEmployeeApplyService.deleteByIdsLogical(canDeleteIds);
+			return result;
+		} else if (canDeleteIds.size()>0 && canDeleteIds.size() < ids.size()) {
+			// 如果部分行可以删除
+			Result result=assetEmployeeApplyService.deleteByIdsLogical(canDeleteIds);
+			if (result.failure()) {
+				return result;
+			} else {
+				return ErrorDesc.success().message("已删除 " + canDeleteIds.size() + " 行，但另有 " + (ids.size() - canDeleteIds.size()) + " 行数据无法删除").messageLevel4Confirm();
+			}
+		} else {
+			// 理论上，这个分支不存在
+			return ErrorDesc.success().message("数据删除未处理");
+		}
 	}
 
 	/**
@@ -129,13 +176,13 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "更新领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.BUSINESS_CODE , value = "业务编号" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.STATUS , value = "办理状态" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class , example = "12"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class , example = "2"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class , example = "1212"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class , example = "0"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.RECORD_TIME , value = "记录时间" , required = false , dataTypeClass=Date.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORIGINATOR_ID , value = "制单人" , required = false , dataTypeClass=String.class),
@@ -156,13 +203,13 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "保存领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.BUSINESS_CODE , value = "业务编号" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.STATUS , value = "办理状态" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class , example = "12"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class , example = "2"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class , example = "1212"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class , example = "0"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.RECORD_TIME , value = "记录时间" , required = false , dataTypeClass=Date.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORIGINATOR_ID , value = "制单人" , required = false , dataTypeClass=String.class),
@@ -227,13 +274,13 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "查询领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.BUSINESS_CODE , value = "业务编号" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.STATUS , value = "办理状态" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class , example = "12"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class , example = "2"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class , example = "1212"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class , example = "0"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.RECORD_TIME , value = "记录时间" , required = false , dataTypeClass=Date.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORIGINATOR_ID , value = "制单人" , required = false , dataTypeClass=String.class),
@@ -255,13 +302,13 @@ public class AssetEmployeeApplyController extends SuperController {
 	*/
 	@ApiOperation(value = "分页查询领用申请")
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "593493381083037696"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.BUSINESS_CODE , value = "业务编号" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.STATUS , value = "办理状态" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class),
-		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NAME , value = "业务名称" , required = false , dataTypeClass=String.class , example = "12"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORG_ID , value = "申请部门" , required = false , dataTypeClass=String.class , example = "2"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.CONTENT , value = "原因" , required = false , dataTypeClass=String.class , example = "1212"),
+		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.APPLY_COUNT , value = "申请数量" , required = false , dataTypeClass=Integer.class , example = "0"),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.RECORD_TIME , value = "记录时间" , required = false , dataTypeClass=Date.class),
 		@ApiImplicitParam(name = AssetEmployeeApplyVOMeta.ORIGINATOR_ID , value = "制单人" , required = false , dataTypeClass=String.class),
@@ -281,6 +328,8 @@ public class AssetEmployeeApplyController extends SuperController {
 		result.success(true).data(list);
 		return result;
 	}
+
+
